+++ source/parlex/include/parlex/state_machine.hpp
@@ -16,11 +16,11 @@ class parser;
 //States from N-a to N-1 are the accept states, where N is states.size() and a is accept_state_count
 class state_machine : public recognizer {
 public:
-	state_machine(std::string id, int acceptStateCount);
-	state_machine(std::string id, int acceptStateCount, filter_function filter);
+	state_machine(std::string id, size_t acceptStateCount);
+	state_machine(std::string id, size_t acceptStateCount, filter_function filter);
 	virtual ~state_machine() = default;
-	void start(details::subjob & sj, int documentPosition) const;
-	void add_transition(int fromState, recognizer const & recognizer, int toState);
+	void start(details::subjob & sj, size_t documentPosition) const;
+	void add_transition(size_t fromState, recognizer const & recognizer, size_t toState);
 	virtual std::string get_id() const final;
 	filter_function get_filter() const;
 
@@ -28,11 +28,11 @@ private:
 	friend class parser;
 
 	std::string const id;
-	int const accept_state_count; //must be greater than 0
-	std::vector<std::map<std::reference_wrapper<recognizer const>, int, details::recognizer_reference_comparer>> states;
+	size_t const accept_state_count; //must be greater than 0
+	std::vector<std::map<std::reference_wrapper<recognizer const>, size_t, details::recognizer_reference_comparer>> states;
 	filter_function const filter;
 
-	void process(details::context_ref const & c, int dfaState) const;
+	void process(details::context_ref const & c, size_t dfaState) const;
 };
 
 }
